+++ internal/client/sender/sender.go
@@ -2,13 +2,12 @@ package sender
 
 import (
 	"bytes"
-	"crypto/hmac"
-	"crypto/sha256"
 	"encoding/base64"
 	"net/http"
 
 	"github.com/ilya372317/must-have-metrics/internal/config"
 	"github.com/ilya372317/must-have-metrics/internal/logger"
+	"github.com/ilya372317/must-have-metrics/internal/signature"
 	"github.com/ilya372317/must-have-metrics/internal/utils/compress"
 )
 
@@ -29,7 +28,7 @@ func SendReport(agentConfig *config.AgentConfig, requestURL, body string) {
 	request.Header.Set("Content-Encoding", "gzip")
 
 	if agentConfig.ShouldSignData() {
-		sign := createSign([]byte(body), agentConfig.SecretKey)
+		sign := signature.CreateSign([]byte(body), agentConfig.SecretKey)
 		encodeSing := base64.StdEncoding.EncodeToString(sign)
 		request.Header.Set("HashSHA256", encodeSing)
 	}
@@ -40,10 +39,4 @@ func SendReport(agentConfig *config.AgentConfig, requestURL, body string) {
 		return
 	}
 	_ = res.Body.Close()
-}
-
-func createSign(body []byte, secretKey string) []byte {
-	h := hmac.New(sha256.New, []byte(secretKey))
-	h.Write(body)
-	return h.Sum(nil)
 }
