+++ internal/handlers/updates.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ilya372317/must-have-metrics/internal/logger"
 	"github.com/ilya372317/must-have-metrics/internal/server/entity"
 	"github.com/ilya372317/must-have-metrics/internal/server/service"
+	"github.com/ilya372317/must-have-metrics/internal/signature"
 )
 
 type BulkUpdateStorage interface {
@@ -25,7 +26,7 @@ type BulkUpdateStorage interface {
 func BulkUpdate(storage BulkUpdateStorage, serverConfig *config.ServerConfig) http.HandlerFunc {
 	return func(writer http.ResponseWriter, request *http.Request) {
 		writer.Header().Set("content-type", "application/json")
-		isSignCorrect, err := isCorrectSigned(serverConfig, request)
+		isSignCorrect, err := signature.IsCorrectSigned(serverConfig, request)
 		if err != nil {
 			http.Error(writer, fmt.Sprintf("failed check sign: %v", err), http.StatusInternalServerError)
 			return
@@ -64,7 +65,7 @@ func BulkUpdate(storage BulkUpdateStorage, serverConfig *config.ServerConfig) ht
 			logger.Log.Warn(err)
 			return
 		}
-		setSign(writer, serverConfig, response)
+		signature.SetSign(writer, serverConfig, response)
 		if _, err = writer.Write(response); err != nil {
 			logger.Log.Warn(err)
 		}
